pyproject.toml
[project]
name = "srl-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "langchain-core>=1.1.0",
    "langchain-openai>=1.1.0",
    "langgraph>=1.0.4",
    "numpy>=2.3.5",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
]
